+++ spec/models/statistics_set_spec.rb
@@ -5,6 +5,13 @@ describe Yt::StatisticsSet do
   subject(:statistics_set) { Yt::StatisticsSet.new data: data }
   let(:value) { 42 }
 
+  describe '#data' do
+    let(:data) { {"key"=>"value"} }
+    specify 'returns the data the statistics set was initialized with' do
+      expect(statistics_set.data).to eq data
+    end
+  end
+
   describe '#view_count' do
     context 'given a video with views' do
       let(:data) { {"viewCount"=>value} }