+++ CHANGELOG.md
@@ -9,6 +9,7 @@ For more information about changelogs, check
 ## 0.14.2 - unreleased
 
 * [FEATURE] Make `Annotation#text` a public method.
+* [FEATURE] Make `data` a public method for Snippet, Status, ContentDetail and StatisticsSet.
 
 ## 0.14.1 - 2015-03-30