+++ CHANGELOG.md
@@ -6,7 +6,7 @@ For more information about changelogs, check
 [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
-## 0.14.2 - unreleased
+## 0.14.2 - 2015-04-08
 
 * [FEATURE] Make `Annotation#text` a public method.
 * [FEATURE] Make `data` a public method for Snippet, Status, ContentDetail and StatisticsSet.